requirements.txt
tgcrypto
pyrogram
python-dotenv
